requirements.txt
streamlit
requests
pandas
transformers
torch
python-dotenv
plotly
wordcloud
Pillow